Zuul depends indirectly on lxml
* modules/zuul/manifests/init.pp: When zuul started using
python-keystoneclient for Swift integration in 36c3fa5 it began
indirectly depending on lxml via keystoneclient. To pip install lxml
you need libxml and libxslt development packages on the system to
provide headers so that the library bindings can be compiled, so
it's easier to just use the python-lxml system packages instead.

Change-Id: Id83ad09f99bcb1606f4037fbb26823a9bd0401ee