Disable arrow alignment puppet lint check
Changes that add or remove parameters often end up shifting all of
the parameter left because of the arrow alignment convention and
check.  This makes it look like more of the code was changed than
really was when reviewing a diff.  Instead, disable the check and
change all arrows to single-space.

Change-Id: I8b945595db479316148c85be6f6c7e12fc7b5fcd