Attempt to add an acceptance test with beaker for puppet-zuul
This patch adds an acceptance test for puppet-zuul. It uses
beaker as framework to perform the test.

This patch should'nt be merged before a solution has been
approved regarding this specification:
https://review.openstack.org/#/c/178887/

This patch could also be abandonned if during discussion around
the mentionned specification another solution than beaker
is selected.

In order to run the test, first install dependencies locally
then start the acceptance test. The default node target
localhost.

$ bundle install
$ bundle exec rspec spec/acceptance

Change-Id: I950e4dd808fda7f3d42b33eeb2d40031e074b2b1