add notification for existing pidfile
The zuul.init and zuul-merger.init puppet module files did not check if pidfiles existed when starting both services. The pidfile's might exist in the event that the zuul-server and merger were killed, which doesn't clean them up. This change checks for whether those files exist and notifies the user to run the stop command which will clean up any residual pid files.

Change-Id: Ie65fdf372f1710ab0c32d03f6b359e1860f566db