Allow configuration of X-Account-Meta-Temp-Url-Key
Normally lib/swift.py would generate and set a new random key each time
zuul restarts. This works fine if swift account is only used by a single
zuul instance. And breaks if two or more instances of zuul would try
to use the same container:

Once the temp key is set on container, only that key can be used with
that container. If new key is set, the old key stops working. So if you
have 2 or more zuul instances all uploading to the same swift container,
only one zuul would win.

So, If account is shared between several zuul instances, the
shared key also has to be used.

Change-Id: Ie192e2ddf2ffece7027ff6fa1e88874784c7dfc7