Ensure zuul gets reinstalled if it's missing
If pip dies in the middle of an uninstall/reinstall cycle, since we only
run pip install on updates to /opt/zuul, we can go days without anything
triggering a reinstall.

Fake this out by running an echo command and marking it as creates. This
will mean puppet will only run it if the file doesn't exist. Then, have
the pip install also subscribe to that echo, so that if it fires, pip
will run an install.

Change-Id: Ida969efc073bf3ca1bc9742bd18f74a1aba65d9c