Zuulv3: Change rewrite rules to get tenant scoped consoles
We recently introduced tenant-scoped resources in zuul-web [1] and
the rewrite rules were not appropriate to redirect to the right place.

Zuul-web is bound to go through more changes in the near future and
will require some amount of refactor to take them into account.

For the time being, this patch assumes a single tenant zuul-web vhost
and adapts the rewrite rules to be under $zuulv3_tenant.

In the future, we need to consider if we want to keep the zuul-web
vhost at the root for a tenant (ex: zuulv3.openstack.org vs
zuulv3.openstack.org/openstack).
If we do, this means that if we want to support a multi-tenant
installation, we need to iterate through a list of domains and tenants
and configure things accordingly.

[1]: https://review.openstack.org/#/c/505452/

Change-Id: Idcb64e76f2466c5adf10315c04ec32db9e593afd