Support updated REST and HTML layout
The previous layout meant we couldn't really do apache offloading of
static content and instead were proxying them to the aiohttp server.

The new update moves api calls below /api. Update the rewrite rules to
deal with this. Also, update the document root to point to where we
deploy the zuul dashboard so that it's picked up by apache.

This also updates the apache config to work for non-whitelabeled deploys
as well.

Depends-On: https://review.openstack.org/556967
Change-Id: I6e662e7bc97de96969828801bcb18c5f354965c5