Add web listen arguments to zuul class
These are used by the zuul.conf template to write the zuul-web
configuration.  The template compares the values to the empty string
to decide if they should be written, since they are undefined, they
are always written, but with no value.  This adds them to the class
with empty-string defaults which should cause them to not appear in
the config file unless needed.

Change-Id: I065d18a0743179cb2c9827a8e6b1b70c169d9199